[*] Renamed use_cases into UseCases to align with naming conventions
