+++ src/UseCases/DataBaseAccess/LoginDataInterface.java
@@ -1,6 +1,5 @@
-package use_cases.DataBaseAccess;
+package UseCases.DataBaseAccess;
 
-import entities.Post;
 import entities.User;
 
 import java.util.ArrayList;
